# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "App AI Gateway API",
    "version": "0.1.0",
    "description": "A multi-tenant, provider-native LLM proxy for mobile applications and trusted server backends."
  },
  "servers": [
    {
      "url": "https://gateway.example.com",
      "description": "Replace with your deployed gateway origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ManagementBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "agw_mgmt_…",
        "description": "An organization-scoped management API key created by an owner or admin."
      },
      "OperatorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agw_operator_auth.session_token",
        "description": "The Better Auth operator session cookie. Admin requests also send x-console-request: 1."
      },
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "A gateway access token, or an application API key for an issuer-less API-key app."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "data": {
                "type": "object",
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "description": "Machine-readable facts about this rejection, present only where the code alone is not actionable. A billing_request_quota_exceeded rejection carries limit, used, and resetAt; an app_rate_limited or app_budget_exhausted rejection carries scope, either user or app."
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/admin/apps/{app}/users": {
      "get": {
        "operationId": "listAppUsers",
        "summary": "List application users",
        "tags": [
          "Admin operations"
        ],
        "security": [
          {
            "OperatorSession": []
          },
          {
            "ManagementBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/apps/{app}/users/{user}": {
      "get": {
        "operationId": "getAppUser",
        "summary": "Get an application user",
        "tags": [
          "Admin operations"
        ],
        "security": [
          {
            "OperatorSession": []
          },
          {
            "ManagementBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "user-123"
            },
            "required": true,
            "name": "user",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}